Implemented testing functions, created new files for training and testing functions for organization
