+++ Proj1.java
@@ -1,34 +1,10 @@
-
 import java.util.Scanner;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
 import java.util.Random;
-import java.util.Arrays;
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.StandardOpenOption;
 
 public class Proj1 {
-
-    //global var for training
-    public static String trainingDataFileName;
-    public static int initWeightsBool;
-    public static int maxEpochs;
-    public static String outputWeightFileName;
-    public static double learningRate;
-    public static double thresholdTheta;
-    public static double thresholdWeightChange;
-
-    //global var for testing
-    public static String trainedWeightsFileName;
-    public static String testingDataFileName;
-    public static String outputTestResultsFileName;
 
     //global data
     public static double[][] global_input_letter_i;
@@ -41,6 +17,7 @@ public class Proj1 {
     public static int outputDimension;
     public static int numberOfLetters;
 
+    public static int initWeightsBool;
 
     public static void main(String[] args) {
         System.out.println("Welcome to our first neural network - A Perceptron Net!");
@@ -54,11 +31,11 @@ public class Proj1 {
         if (validateInteger(scanner)) return;
         int action = scanner.nextInt();
         while (action != 3) {
-
+                
             if (action == 1) {
-                trainingSpecs(scanner);
+                training.trainingSpecs(scanner);
             } else if (action == 2) {
-                testingSpecs(scanner);
+                testing.testingSpecs(scanner);
             }
             System.out.println("Welcome to our first neural network - A Perceptron Net!");
             System.out.println("1) Enter 1 to train the net on a data file");
@@ -68,97 +45,9 @@ public class Proj1 {
         }
         System.out.println("Thank you for using the Net. Come back soon!");
 
-
     }
 
-    public static double calculateY_in( double [] weights, double[] inputs) {
-        double linearCombination = 0;
-        //Add the bias
-        for(int idx = 0; idx < inputs.length; idx++){
-            double product = (inputs[idx] * weights[idx]);
-            linearCombination += product;
-        }
-        return linearCombination;
-    }
-
-    public static int classifyLinearCombination(double linearCombination, double threshold) {
-        if (linearCombination > threshold) {
-            return 1;
-        } else if (linearCombination < (-1 * threshold)) {
-            return -1;
-        } else {
-            return 0;
-        }
-    }
-
-    public static void trainingSpecs(Scanner scanner){
-
-        System.out.println("Enter the training data file name:");
-        scanner.nextLine(); //get rid of newline
-        trainingDataFileName = scanner.nextLine();
-
-        System.out.println("Enter 0 to initialize weights to 0, enter 1 to initialize weights to random values between -0.5 and 0.5:");
-        initWeightsBool = scanner.nextInt();
-
-        System.out.println("Enter the maximum number of training epochs:");
-        maxEpochs = scanner.nextInt();
-
-        System.out.println("Enter a file name to save the trained weight values:");
-        scanner.nextLine();
-        outputWeightFileName = scanner.nextLine();
-
-        System.out.println("Enter the learning rate alpha from 0 to 1 but not including 0:");
-        learningRate = scanner.nextDouble();
-
-        System.out.println("Enter the threshold theta:");
-        thresholdTheta = scanner.nextDouble();
-
-        System.out.println("Enter the threshold to be used for measuring weight changes:");
-        thresholdWeightChange = scanner.nextDouble();
-
-        //scanner.close(); // close scnner
-
-        //call data read method
-        readData(trainingDataFileName);
-
-        //train
-        int epochCountAfterTraining = trainAlgorithm(maxEpochs, outputWeightFileName, learningRate, thresholdTheta, thresholdWeightChange, global_weights_j_i, global_input_letter_i, global_target_letter_j, outputDimension, colDimension*outputDimension);
-        if(epochCountAfterTraining == 0){
-            System.out.println("Training did not converge in less than the max amount of epochs.");
-        }
-        else{
-            System.out.println("Training converged after "+epochCountAfterTraining+ " epochs");
-        }
-
-        //save weights to file
-
-
-    }
-
-    public static void testingSpecs(Scanner scanner){
-
-        System.out.println("Enter the trained net weight file name:");
-        scanner.nextLine(); //get rid of newline
-        trainedWeightsFileName = scanner.nextLine();
-
-        System.out.println("Enter the testing/deploying dataset file name:");
-        testingDataFileName = scanner.nextLine();
-
-        System.out.println("Enter a file name to save the testing/deploying results:");
-        outputTestResultsFileName = scanner.nextLine();
-
-        //scanner.close();
-
-        //call data read method
-        readData(testingDataFileName);
-
-        //call another method to read weights
-
-        //test and save outputs to a file
-
-    }
-
-    private static boolean validateInteger(Scanner myScanner) {
+    public static boolean validateInteger(Scanner myScanner) {
         if (!myScanner.hasNextInt()) {
             System.out.println("Invalid x: Please enter an integer.");
             myScanner.close();
@@ -166,6 +55,7 @@ public class Proj1 {
         }
         return false;
     }
+
     /**
      * this function reads the input from a file into input_letter_i and target_letter_j, which encases all the input for each letter
      * */
@@ -270,88 +160,4 @@ public class Proj1 {
         System.out.println("Number of Rows: " + count);
     }
 
-        //write algorithm
-    public static int trainAlgorithm(int maxEpochs, String outputFileName,
-                                     double learningRate, double thresholdTheta,
-                                     double thresholdWeightChange, double[][] weights,
-                                     double[][] inputSamples, int[][] targetOutputs,
-                                     int numOutputNeurons, int numInputFeatures) {
-
-        boolean hasConverged = false;
-        int epochCount = 0;
-        double[] weightChangeAmounts = new double[numInputFeatures + 1];  // Track weight changes
-
-        while (!hasConverged && epochCount < maxEpochs) {
-            boolean weightUpdated = false;
-            Arrays.fill(weightChangeAmounts, 0.0);  // Reset weight changes for this epoch
-
-            for (int outputNeuron = 0; outputNeuron < numOutputNeurons; outputNeuron++) {
-                for (int sampleIdx = 0; sampleIdx < targetOutputs.length; sampleIdx++) {
-
-                    int expectedOutput = targetOutputs[sampleIdx][outputNeuron];  // Get correct target
-                    if (expectedOutput != -1 && expectedOutput != 1) {
-                        System.err.println("Error: Target output must be -1 or 1, but got: " + expectedOutput);
-                        continue; // Skip invalid data
-                    }
-
-                    //Compute perceptron output
-                    double netInput = calculateY_in(weights[outputNeuron], inputSamples[sampleIdx]);
-                    int predictedOutput = classifyLinearCombination(netInput, thresholdTheta);
-
-                    //Check if weights need updating
-                    if (expectedOutput != predictedOutput) {
-                        for (int inputIdx = 0; inputIdx < numInputFeatures; inputIdx++) {
-                            double previousWeight = weights[outputNeuron][inputIdx];
-                            weights[outputNeuron][inputIdx] += learningRate * expectedOutput * inputSamples[sampleIdx][inputIdx];
-                            weightChangeAmounts[inputIdx] = Math.abs(weights[outputNeuron][inputIdx] - previousWeight); // Track weight change
-                        }
-
-                        weightUpdated = true;
-                    }
-                }
-            }
-
-            //Check for convergence
-            double maxWeightChange = findMaxWeightChange(weightChangeAmounts);
-            if (maxWeightChange < thresholdWeightChange) {
-                hasConverged = true;
-            } else if (!weightUpdated) {
-                hasConverged = true;
-            } else {
-                epochCount++;
-            }
-        }
-
-        saveWeightsToFile(outputFileName, epochCount, weights);
-
-        return (epochCount == maxEpochs) ? 0 : epochCount;
-    }
-
-
-    public static double findMaxWeightChange(double[] weightChanges) {
-        double maxWeightChange = Math.abs(weightChanges[0]);  //Ensure absolute values are considered
-        for (int i = 1; i < weightChanges.length; i++) {  //Start from 1 since 0 is initialized
-            if (Math.abs(weightChanges[i]) > maxWeightChange) {
-                maxWeightChange = Math.abs(weightChanges[i]);
-            }
-        }
-        return maxWeightChange;
-    }
-
-    public static void saveWeightsToFile(String filename, int epoch, double[][] weights) {
-        try {
-            Path path = Paths.get(filename);
-            StringBuilder weightData = new StringBuilder();
-            weightData.append("Epoch ").append(epoch).append(System.lineSeparator());
-
-            for (double[] outputNeuronWeights : weights) {
-                weightData.append(Arrays.toString(outputNeuronWeights)).append(System.lineSeparator());
-            }
-
-            Files.write(path, weightData.toString().getBytes(StandardCharsets.UTF_8),
-                    StandardOpenOption.APPEND, StandardOpenOption.CREATE);
-        } catch (IOException e) {
-            System.err.println("Error writing weights to file: " + e.getMessage());
-        }
-    }
 }
